[FLINK-13488][tests] Harden ConnectedComponents E2E test
By default the tests starts 25 TMs with a single slot each. This is not
sustainable on Travis CI. This commit changes the test so that it only starts 2
TMs that each offer 13 slots by default.

Run 'set -Eexuo pipefail' at the beginning of the test as recommended by the
README.md.